Better support for multi-line conditional predicate
